Fix staking minimum age detection
nBlockTime is not a reliable source for determining when a generated block was created, as it is often '0'.
This fix adds extra logic to guard against this by using GetAdjustedTime() as a fallback option.

Additional debugging output has also been written that can be enabled with the `-debug` or `debug=1`
runtime options.