[Qt] Add a timer to update the staking icon and weights
To ensure the staking icon is updated in a timely manner, add a
timer to the process that will update it as needed.

Previous behavior seemed inconsistant, where the icon's state
wasn't being updated after changes to the wallet's locking status
or when the wallet ran out of mature inputs to stake with.